Renamed extra to utils folder
